fix delete taxonomy

+++ war/src/main/java/co/xarx/trix/eventhandler/StationEventHandler.java
@@ -186,12 +186,21 @@ public class StationEventHandler {
 			stationPerspectiveRepository.delete(stationsPerspectives);
 
 			Taxonomy taxonomy = taxonomyRepository.findOne(station.categoriesTaxonomyId);
-			taxonomy.owningStation = null;
+			Taxonomy taxonomyTags = taxonomyRepository.findOne(station.tagsTaxonomyId);
+			Taxonomy taxonomyAuthors = taxonomyRepository.findTypeAByStation(station);
 
-			taxonomyRepository.save(taxonomy);
-
-			taxonomyEventHandler.handleBeforeDelete(taxonomy);
-			taxonomyRepository.delete(taxonomy);
+			if (taxonomy != null) {
+				taxonomyEventHandler.handleBeforeDelete(taxonomy);
+				taxonomyRepository.delete(taxonomy);
+			}
+			if (taxonomyTags != null) {
+				taxonomyEventHandler.handleBeforeDelete(taxonomyTags);
+				taxonomyRepository.delete(taxonomyTags);
+			}
+			if (taxonomyAuthors != null) {
+				taxonomyEventHandler.handleBeforeDelete(taxonomyAuthors);
+				taxonomyRepository.delete(taxonomyAuthors);
+			}
 
 			List<StationRole> stationsRoles = personStationRolesRepository.findByStation(station);
 			if(stationsRoles != null && stationsRoles.size() > 0){
